% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbl2helpers.R
\name{add_variable_padding}
\alias{add_variable_padding}
\title{Add padding to exported table variables}
\usage{
add_variable_padding(df)
}
\description{
Add padding to exported table variables
}
\keyword{internal}
