% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbl2helpers.R
\name{apply_zebra_stripes_ft}
\alias{apply_zebra_stripes_ft}
\title{Apply zebra stripes with proper variable group detection for indented tables}
\usage{
apply_zebra_stripes_ft(ft, df, var_groups)
}
\description{
Apply zebra stripes with proper variable group detection for indented tables
}
\keyword{internal}
