% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbl2helpers.R
\name{bold_pvalues_ft}
\alias{bold_pvalues_ft}
\title{Bold significant p-values in DOCX}
\usage{
bold_pvalues_ft(ft, df, sig_threshold = 0.05)
}
\description{
Bold significant p-values in DOCX
}
\keyword{internal}
