% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comphelpers.R
\name{build_comparison_table}
\alias{build_comparison_table}
\title{Build comprehensive comparison table}
\usage{
build_comparison_table(comparison, model_type)
}
\description{
Build comprehensive comparison table
}
\keyword{internal}
