% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/desctbl.R
\name{calc_continuous_stat}
\alias{calc_continuous_stat}
\title{Calculate continuous statistic}
\usage{
calc_continuous_stat(x, stat_type, digits)
}
\description{
Calculate continuous statistic
}
\keyword{internal}
