% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foresthelpers.R
\name{calculate_forest_layout}
\alias{calculate_forest_layout}
\title{Calculate table layout for forest plots}
\usage{
calculate_forest_layout(
  to_show_exp_clean,
  show_n,
  show_events,
  indent_groups,
  condense_table,
  effect_label,
  ref_label,
  font_size,
  tbl_width = 0.6,
  rangeb,
  center_padding
)
}
\description{
Calculate table layout for forest plots
}
\keyword{internal}
