% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comphelpers.R
\name{calculate_model_scores}
\alias{calculate_model_scores}
\title{Calculate FastFit scores for model comparison}
\usage{
calculate_model_scores(comparison, model_type, scoring_weights = NULL)
}
\description{
Calculate FastFit scores for model comparison
}
\keyword{internal}
