% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbl2helpers.R
\name{calculate_table_width}
\alias{calculate_table_width}
\title{Calculate table width based on paper size and orientation}
\usage{
calculate_table_width(paper, orientation)
}
\description{
Calculate table width based on paper size and orientation
}
\keyword{internal}
