% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbl2helpers.R
\name{check_latex}
\alias{check_latex}
\title{Export Helper Functions}
\usage{
check_latex()
}
\description{
Various helper functions for exporting fastfit tables to PDF, LaTeX, HTML, etc.
}
\details{
Check if LaTeX is available
}
\keyword{internal}
