% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbl2helpers.R
\name{condense_binary}
\alias{condense_binary}
\title{Condense binary categorical variable rows}
\usage{
condense_binary(data, start_row, end_row, rows_to_keep, is_descriptive)
}
\description{
Condense binary categorical variable rows
}
\keyword{internal}
