% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbl2helpers.R
\name{condense_continuous}
\alias{condense_continuous}
\title{Condense continuous variable rows}
\usage{
condense_continuous(data, start_row, end_row, rows_to_keep, is_descriptive)
}
\description{
Condense continuous variable rows
}
\keyword{internal}
