% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbl2helpers.R
\name{condense_table_rows}
\alias{condense_table_rows}
\title{Condense table rows for more compact display}
\usage{
condense_table_rows(df, indent_groups = TRUE)
}
\description{
Condense table rows for more compact display
}
\keyword{internal}
