% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m2dt.R
\name{detect_model_type}
\alias{detect_model_type}
\title{Detect if model is univariable or multivariable}
\usage{
detect_model_type(model)
}
\description{
Detect if model is univariable or multivariable
}
\keyword{internal}
