% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbl2helpers.R
\name{determine_alignment}
\alias{determine_alignment}
\title{Determine alignment for exported tables}
\usage{
determine_alignment(df)
}
\description{
Determine alignment for exported tables
}
\keyword{internal}
