% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastfit.R
\name{determine_effect_type}
\alias{determine_effect_type}
\title{Helper function to determine effect type - EXTRACTED FOR CLARITY}
\usage{
determine_effect_type(uni_raw, multi_raw, exponentiate, adjusted = FALSE)
}
\description{
Helper function to determine effect type - EXTRACTED FOR CLARITY
}
\keyword{internal}
