% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbl2helpers.R
\name{format_column_headers}
\alias{format_column_headers}
\title{Apply formatting to column headers in exported tables}
\usage{
format_column_headers(col_names, add_header_space = TRUE)
}
\description{
Apply formatting to column headers in exported tables
}
\keyword{internal}
