% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbl2helpers.R
\name{format_column_headers_with_n_html}
\alias{format_column_headers_with_n_html}
\title{Format column headers with n counts (HTML)}
\usage{
format_column_headers_with_n_html(col_names, n_row_data)
}
\description{
Format column headers with n counts (HTML)
}
\keyword{internal}
