% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbl2helpers.R
\name{format_column_headers_with_n_tex}
\alias{format_column_headers_with_n_tex}
\title{Format column headers with n counts (TeX)}
\usage{
format_column_headers_with_n_tex(col_names, n_row_data)
}
\description{
Format column headers with n counts (TeX)
}
\keyword{internal}
