% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastfit.R
\name{format_fastfit_table}
\alias{format_fastfit_table}
\title{Format fastfit table output}
\usage{
format_fastfit_table(
  uni_results,
  multi_results,
  predictors,
  columns,
  metrics,
  model_type
)
}
\description{
Format fastfit table output
}
\keyword{internal}
