% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comphelpers.R
\name{format_model_comparison}
\alias{format_model_comparison}
\title{Format model comparison table}
\usage{
format_model_comparison(comparison)
}
\description{
Format model comparison table
}
\keyword{internal}
