% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deschelpers.R
\name{format_pvalues_desctbl}
\alias{format_pvalues_desctbl}
\title{Format p-values in result table}
\usage{
format_pvalues_desctbl(result, digits_p)
}
\description{
Format p-values in result table
}
\keyword{internal}
