% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbl2helpers.R
\name{format_pvalues_export_html}
\alias{format_pvalues_export_html}
\title{Format p-values for exported tables (HTML)}
\usage{
format_pvalues_export_html(df, p_threshold = 0.05)
}
\description{
Format p-values for exported tables (HTML)
}
\keyword{internal}
