% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbl2helpers.R
\name{format_pvalues_export_tex}
\alias{format_pvalues_export_tex}
\title{Format p-values for exported tables}
\usage{
format_pvalues_export_tex(df, p_threshold = 0.05)
}
\description{
Format p-values for exported tables
}
\keyword{internal}
