% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/desctbl.R
\name{format_survival_stat}
\alias{format_survival_stat}
\title{Format survival statistic}
\usage{
format_survival_stat(median, lower, upper, digits)
}
\description{
Format survival statistic
}
\keyword{internal}
