% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbl2helpers.R
\name{get_paper_settings}
\alias{get_paper_settings}
\title{Get paper size for PDF/LaTeX export}
\usage{
get_paper_settings(paper, margins = NULL)
}
\description{
Get paper size for PDF/LaTeX export
}
\keyword{internal}
