% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbl2helpers.R
\name{identify_variable_groups}
\alias{identify_variable_groups}
\title{Identify variable groups before indentation}
\usage{
identify_variable_groups(df)
}
\description{
Identify variable groups before indentation
}
\keyword{internal}
