% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{or_or}
\alias{or_or}
\alias{\%||\%}
\title{Null coalescing operator}
\usage{
a \%||\% b
}
\arguments{
\item{a}{Left side value}

\item{b}{Right side value (default)}
}
\value{
a if not NULL, otherwise b
}
\description{
Returns the left-hand side if not NULL, otherwise returns the right-hand side.
}
\keyword{internal}
