% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/desctbl.R
\name{perform_categorical_test}
\alias{perform_categorical_test}
\title{Perform categorical variable test}
\usage{
perform_categorical_test(data, var, group_var, test_type, ...)
}
\description{
Perform categorical variable test
}
\keyword{internal}
