% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deschelpers.R
\name{perform_continuous_test}
\alias{perform_continuous_test}
\title{Perform continuous variable test}
\usage{
perform_continuous_test(data, var, group_var, test_type, stat_type = NULL, ...)
}
\description{
Perform continuous variable test
}
\keyword{internal}
