% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbl2docx.R
\name{print.tbl2docx_result}
\alias{print.tbl2docx_result}
\title{Print method for tbl2docx results}
\usage{
\method{print}{tbl2docx_result}(x, ...)
}
\description{
Print method for tbl2docx results
}
\keyword{internal}
