% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uscreen.R
\name{print.uscreen_result}
\alias{print.uscreen_result}
\title{Print method for uscreen results}
\usage{
\method{print}{uscreen_result}(x, n = 20, ...)
}
\description{
Print method for uscreen results
}
\keyword{internal}
