% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/desctbl.R
\name{process_continuous}
\alias{process_continuous}
\title{Process continuous variable}
\usage{
process_continuous(
  data,
  var,
  var_label,
  group_var,
  stats,
  digits,
  na_include,
  na_label,
  test,
  test_type,
  total,
  total_label,
  ...
)
}
\description{
Process continuous variable
}
\keyword{internal}
