% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/desctbl.R
\name{process_survival}
\alias{process_survival}
\title{Process survival variable - matching column structure}
\usage{
process_survival(
  data,
  var,
  var_label,
  group_var,
  digits,
  test,
  total,
  total_label,
  ...
)
}
\description{
Process survival variable - matching column structure
}
\keyword{internal}
