% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deschelpers.R
\name{process_variable}
\alias{process_variable}
\title{Process variable wrapper - returns both formatted and raw}
\usage{
process_variable(
  data,
  var,
  group_var = NULL,
  stats_continuous,
  stats_categorical,
  digits,
  na_include,
  na_label,
  test,
  test_continuous,
  test_categorical,
  total,
  total_label,
  labels,
  na_percent,
  ...
)
}
\description{
Process variable wrapper - returns both formatted and raw
}
\keyword{internal}
