% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbl2helpers.R
\name{process_variable_block}
\alias{process_variable_block}
\title{Process a block of rows for a single variable}
\usage{
process_variable_block(data, start_row, end_row, rows_to_keep, is_descriptive)
}
\description{
Process a block of rows for a single variable
}
\keyword{internal}
