% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastfit.R
\name{format_fastfit_combined}
\alias{format_fastfit_combined}
\title{Format combined fastfit output from formatted tables}
\usage{
format_fastfit_combined(
  uni_formatted,
  multi_formatted,
  uni_raw,
  multi_raw,
  predictors,
  columns,
  metrics,
  show_n_events,
  var_labels,
  exponentiate = NULL
)
}
\description{
Format combined fastfit output from formatted tables
}
\keyword{internal}
