% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comphelpers.R
\name{format_model_comparison}
\alias{format_model_comparison}
\title{Format comparison table for display}
\usage{
format_model_comparison(dt)
}
\description{
Format comparison table for display
}
\keyword{internal}
