% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deschelpers.R
\name{process_survival}
\alias{process_survival}
\title{Process survival variable}
\usage{
process_survival(
  data,
  var,
  var_label,
  group_var,
  digits,
  test,
  total,
  total_label,
  na_include = FALSE,
  na_label = "Unknown",
  ...
)
}
\description{
Process survival variable
}
\keyword{internal}
