% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/desctbl.R
\name{process_variable}
\alias{process_variable}
\title{Process single variable for descriptive statistics}
\usage{
process_variable(
  data,
  var,
  group_var = NULL,
  stats_continuous,
  stats_categorical,
  digits,
  na_include,
  na_label,
  test,
  test_continuous,
  test_categorical,
  total,
  total_label,
  var_labels,
  na_percent,
  ...
)
}
\description{
Process single variable for descriptive statistics
}
\keyword{internal}
